+++ packages/xchain-thorchain/src/util.ts
@@ -1,3 +1,4 @@
+import { cosmosclient, proto } from '@cosmos-client/core'
 import { Address, Balance, FeeType, Fees, Network, TxHash, TxType, singleFee } from '@thorswap-lib/xchain-client'
 import { CosmosSDKClient, TxLog } from '@thorswap-lib/xchain-cosmos'
 import {
@@ -5,18 +6,18 @@ import {
   AssetRuneNative,
   BaseAmount,
   Chain,
+  assetAmount,
   assetFromString,
+  assetToBase,
   assetToString,
   baseAmount,
 } from '@thorswap-lib/xchain-util'
 import axios from 'axios'
-import { AccAddress, Msg, codec } from 'cosmos-client'
-import { StdTxFee } from 'cosmos-client/api'
-import { StdTx } from 'cosmos-client/x/auth'
-import { MsgMultiSend, MsgSend } from 'cosmos-client/x/bank'
+import * as bech32Buffer from 'bech32-buffer'
 
-import { ClientUrl, ExplorerUrl, ExplorerUrls, TxData } from './types'
-import { MsgNativeTx, ThorchainDepositResponse } from './types/messages'
+import { ChainId, ChainIds, ClientUrl, ExplorerUrl, ExplorerUrls, NodeInfoResponse, TxData } from './types'
+import { MsgNativeTx } from './types/messages'
+import types from './types/proto/MsgCompiled'
 
 export const DECIMAL = 8
 export const DEFAULT_GAS_VALUE = '500000000'
@@ -24,7 +25,7 @@ export const DEPOSIT_GAS_VALUE = '500000000'
 export const MAX_TX_COUNT = 100
 
 export const THORCHAIN_MAINNET_CHAIN_ID = 'thorchain-mainnet-v1'
-export const THORCHAIN_STAGENET_CHAIN_ID = 'thorchain-stagenet-v1'
+export const THORCHAIN_STAGENET_CHAIN_ID = 'thorchain-stagenet-v2'
 export const THORCHAIN_TESTNET_CHAIN_ID = 'thorchain-testnet-v2'
 
 /**
@@ -65,28 +66,13 @@ export const getDenomWithChain = (asset: Asset): string => {
  */
 export const getAsset = (denom: string): Asset | null => {
   if (denom === getDenom(AssetRuneNative)) return AssetRuneNative
+
+  // use lowercase symbol for synth
+  if (denom.includes('/')) {
+    return assetFromString(`${Chain.THORChain}.${denom.toLowerCase()}`)
+  }
   return assetFromString(`${Chain.THORChain}.${denom.toUpperCase()}`)
 }
-
-/**
- * Type guard for MsgSend
- *
- * @param {Msg} msg
- * @returns {boolean} `true` or `false`.
- */
-export const isMsgSend = (msg: Msg): msg is MsgSend =>
-  (msg as MsgSend)?.amount !== undefined &&
-  (msg as MsgSend)?.from_address !== undefined &&
-  (msg as MsgSend)?.to_address !== undefined
-
-/**
- * Type guard for MsgMultiSend
- *
- * @param {Msg} msg
- * @returns {boolean} `true` or `false`.
- */
-export const isMsgMultiSend = (msg: Msg): msg is MsgMultiSend =>
-  (msg as MsgMultiSend)?.inputs !== undefined && (msg as MsgMultiSend)?.outputs !== undefined
 
 /**
  * Response guard for transaction broadcast
@@ -123,23 +109,17 @@ export const getPrefix = (network: Network, isStagenet = false) => {
   }
 }
 
+export const registerDespositCodecs = async (): Promise<void> => {
+  cosmosclient.codec.register('/types.MsgDeposit', types.types.MsgDeposit)
+}
+
 /**
  * Register Codecs based on the prefix.
  *
  * @param {string} prefix
  */
-export const registerCodecs = (prefix: string): void => {
-  codec.registerCodec('thorchain/MsgSend', MsgSend, MsgSend.fromJSON)
-  codec.registerCodec('thorchain/MsgMultiSend', MsgMultiSend, MsgMultiSend.fromJSON)
-
-  AccAddress.setBech32Prefix(
-    prefix,
-    prefix + 'pub',
-    prefix + 'valoper',
-    prefix + 'valoperpub',
-    prefix + 'valcons',
-    prefix + 'valconspub',
-  )
+export const registerSendCodecs = async (): Promise<void> => {
+  cosmosclient.codec.register('/types.MsgSend', types.types.MsgSend)
 }
 
 /**
@@ -194,7 +174,7 @@ export const getDepositTxDataFromLogs = (logs: TxLog[], address: Address): TxDat
  * @returns {Fees} The default fee.
  */
 export const getDefaultFees = (): Fees => {
-  const fee = baseAmount(DEFAULT_GAS_VALUE, DECIMAL)
+  const fee = assetToBase(assetAmount(0.02 /* 0.02 RUNE */, DECIMAL))
   return singleFee(FeeType.FlatFee, fee)
 }
 
@@ -210,38 +190,166 @@ export const getTxType = (txData: string, encoding: 'base64' | 'hex'): string =>
 }
 
 /**
+ * Helper to get THORChain's chain id
+ * @param {string} nodeUrl THORNode url
+ */
+export const getChainId = async (nodeUrl: string): Promise<string> => {
+  const { data } = await axios.get<NodeInfoResponse>(`${nodeUrl}/cosmos/base/tendermint/v1beta1/node_info`)
+  return data?.default_node_info?.network || Promise.reject('Could not parse chain id')
+}
+
+/**
+ * Helper to get all THORChain's chain id
+ * @param {ClientUrl} client urls (use `getDefaultClientUrl()` if you don't need to use custom urls)
+ */
+export const getChainIds = async (client: ClientUrl): Promise<ChainIds> => {
+  return Promise.all([getChainId(client[Network.Testnet].node), getChainId(client[Network.Mainnet].node)]).then(
+    ([testnetId, mainnetId]) => ({
+      testnet: testnetId,
+      mainnet: mainnetId,
+    }),
+  )
+}
+
+/**
  * Structure StdTx from MsgNativeTx.
  *
- * @param {string} txId The transaction id.
+ * @param {MsgNativeTx} msgNativeTx Msg of type `MsgNativeTx`.
+ * @param {string} nodeUrl Node url
+ * @param {chainId} ChainId Chain id of the network
+ *
  * @returns {Tx} The transaction details of the given transaction id.
  *
  * @throws {"Invalid client url"} Thrown if the client url is an invalid one.
  */
-export const buildDepositTx = async (msgNativeTx: MsgNativeTx, nodeUrl: string, chainId: string): Promise<StdTx> => {
-  const response: ThorchainDepositResponse = (
-    await axios.post(`${nodeUrl}/thorchain/deposit`, {
-      coins: msgNativeTx.coins,
-      memo: msgNativeTx.memo,
-      base_req: {
-        chain_id: chainId,
-        from: msgNativeTx.signer,
+export const buildDepositTx = async ({
+  msgNativeTx,
+  nodeUrl,
+  chainId,
+}: {
+  msgNativeTx: MsgNativeTx
+  nodeUrl: string
+  chainId: ChainId
+}): Promise<proto.cosmos.tx.v1beta1.TxBody> => {
+  const networkChainId = await getChainId(nodeUrl)
+  if (!networkChainId || chainId !== networkChainId) {
+    throw new Error(`Invalid network (asked: ${chainId} / returned: ${networkChainId}`)
+  }
+
+  const signerAddr = msgNativeTx.signer.toString()
+  const signerDecoded = bech32Buffer.decode(signerAddr)
+
+  const msgDepositObj = {
+    coins: msgNativeTx.coins,
+    memo: msgNativeTx.memo,
+    signer: signerDecoded.data,
+  }
+
+  const depositMsg = types.types.MsgDeposit.fromObject(msgDepositObj)
+
+  return new proto.cosmos.tx.v1beta1.TxBody({
+    messages: [cosmosclient.codec.packAny(depositMsg)],
+    memo: msgNativeTx.memo,
+  })
+}
+
+/**
+ * Structure a MsgSend
+ *
+ * @param fromAddress - required, from address string
+ * @param toAddress - required, to address string
+ * @param assetAmount - required, asset amount string (e.g. "10000")
+ * @param assetDenom - required, asset denom string (e.g. "rune")
+ * @param memo - optional, memo string
+ *
+ * @returns
+ */
+export const buildTransferTx = async ({
+  fromAddress,
+  toAddress,
+  assetAmount,
+  assetDenom,
+  memo = '',
+  nodeUrl,
+  chainId,
+}: {
+  fromAddress: Address
+  toAddress: Address
+  assetAmount: BaseAmount
+  assetDenom: string
+  memo?: string
+  nodeUrl: string
+  chainId: ChainId
+}): Promise<proto.cosmos.tx.v1beta1.TxBody> => {
+  const networkChainId = await getChainId(nodeUrl)
+  if (!networkChainId || chainId !== networkChainId) {
+    throw new Error(`Invalid network (asked: ${chainId} / returned: ${networkChainId}`)
+  }
+
+  const fromDecoded = bech32Buffer.decode(fromAddress)
+  const toDecoded = bech32Buffer.decode(toAddress)
+
+  const transferObj = {
+    fromAddress: fromDecoded.data,
+    toAddress: toDecoded.data,
+    amount: [
+      {
+        amount: assetAmount.amount().toString(),
+        denom: assetDenom,
       },
-    })
-  ).data
+    ],
+  }
 
-  if (!response || !response.value) throw new Error('Invalid client url')
+  const transferMsg = types.types.MsgSend.fromObject(transferObj)
 
-  const fee: StdTxFee = response.value?.fee ?? { amount: [] }
+  return new proto.cosmos.tx.v1beta1.TxBody({
+    messages: [cosmosclient.codec.packAny(transferMsg)],
+    memo,
+  })
+}
 
-  const unsignedStdTx = StdTx.fromJSON({
-    msg: response.value.msg,
-    // override fee
-    fee: { ...fee, gas: DEPOSIT_GAS_VALUE },
-    signatures: [],
-    memo: '',
+/**
+ * Builds final unsigned TX
+ *
+ * @param cosmosSdk - CosmosSDK
+ * @param txBody - txBody with encoded Msgs
+ * @param signerPubkey - signerPubkey string
+ * @param sequence - account sequence
+ * @param gasLimit - transaction gas limit
+ * @returns
+ */
+export const buildUnsignedTx = ({
+  cosmosSdk,
+  txBody,
+  signerPubkey,
+  sequence,
+  gasLimit,
+}: {
+  cosmosSdk: cosmosclient.CosmosSDK
+  txBody: proto.cosmos.tx.v1beta1.TxBody
+  signerPubkey: proto.google.protobuf.Any
+  sequence: cosmosclient.Long
+  gasLimit: string
+}): cosmosclient.TxBuilder => {
+  const authInfo = new proto.cosmos.tx.v1beta1.AuthInfo({
+    signer_infos: [
+      {
+        public_key: signerPubkey,
+        mode_info: {
+          single: {
+            mode: proto.cosmos.tx.signing.v1beta1.SignMode.SIGN_MODE_DIRECT,
+          },
+        },
+        sequence: sequence,
+      },
+    ],
+    fee: {
+      amount: null,
+      gas_limit: cosmosclient.Long.fromString(gasLimit),
+    },
   })
 
-  return unsignedStdTx
+  return new cosmosclient.TxBuilder(cosmosSdk, txBody, authInfo)
 }
 
 /**
@@ -263,11 +371,14 @@ export const getBalance = async ({
   cosmosClient: CosmosSDKClient
 }): Promise<Balance[]> => {
   const balances = await cosmosClient.getBalance(address)
+
   return balances
-    .map((balance) => ({
-      asset: (balance.denom && getAsset(balance.denom)) || AssetRuneNative,
-      amount: baseAmount(balance.amount, DECIMAL),
-    }))
+    .map((balance) => {
+      return {
+        asset: (balance.denom && getAsset(balance.denom)) || AssetRuneNative,
+        amount: baseAmount(balance.amount, DECIMAL),
+      }
+    })
     .filter(
       (balance) => !assets || assets.filter((asset) => assetToString(balance.asset) === assetToString(asset)).length,
     )
@@ -283,7 +394,7 @@ export const getDefaultClientUrl = (isStagenet = false): ClientUrl => {
     return {
       [Network.Testnet]: {
         node: 'https://testnet.thornode.thorchain.info',
-        rpc: 'https://testnet.rpc.thorchain.info',
+        rpc: 'https://testnet-rpc.ninerealms.com',
       },
       [Network.Mainnet]: {
         node: 'https://stagenet-thornode.ninerealms.com',
@@ -295,11 +406,11 @@ export const getDefaultClientUrl = (isStagenet = false): ClientUrl => {
   return {
     [Network.Testnet]: {
       node: 'https://testnet.thornode.thorchain.info',
-      rpc: 'https://testnet.rpc.thorchain.info',
+      rpc: 'https://testnet-rpc.ninerealms.com',
     },
     [Network.Mainnet]: {
-      node: 'https://thornode.thorchain.info',
-      rpc: 'https://rpc.thorchain.info',
+      node: 'https://thornode.ninerealms.com',
+      rpc: 'https://rpc.ninerealms.com',
     },
   }
 }
@@ -408,3 +519,5 @@ export const getExplorerTxUrl = ({
       return `${url}?network=testnet`
   }
 }
+
+export const isAssetRuneNative = (asset: Asset): boolean => assetToString(asset) === assetToString(AssetRuneNative)